+++ site/src/pages/UserSettingsPage/AccountPage/AccountPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react"
 import { Section } from "../../../components/SettingsLayout/Section"
-import { AccountForm } from "../../../components/SettingsAccountForm/SettingsAccountForm"
+import { AccountForm } from "./SettingsAccountForm/SettingsAccountForm"
 import { useAuth } from "components/AuthProvider/AuthProvider"
 import { useMe } from "hooks/useMe"
 import { usePermissions } from "hooks/usePermissions"
